add react router and error page
