+++ web/package.json
@@ -10,6 +10,7 @@
     "react": "16.7.0",
     "react-bootstrap": "^0.32.1",
     "react-dom": "16.7.0",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3"
   },
   "scripts": {
