+++ web/src/App.css
@@ -26,3 +26,22 @@
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
+
+.oops {
+  display: inline-block;
+  background: #00AFF9 no-repeat;
+  height: 100vh;
+  width: 100vw;
+  margin: 0;
+  color: white;
+}
+
+.whoops {
+  margin: .8em 3rem;
+  font: 4em ;
+}
+.wrong {
+  display: inline-block;
+  margin: .2em 3rem;
+  font: 2em ;
+}
