+++ web/src/components/page-not-found.js
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+
+export default class PageNotFound extends Component {
+  render() {
+    return (
+      <div className="oops">
+        <h1 className="whoops">Whoops!</h1>
+        <p className="wrong">Something went wrong</p>
+      </div>
+    );
+  }
+}
